+++ modules/bridge/src/main/ts/ephox/bridge/components/dialog/Iframe.ts
@@ -6,18 +6,18 @@ import { FormComponentWithLabel, formComponentWithLabelFields, FormComponentWith
 export interface IframeSpec extends FormComponentWithLabelSpec {
   type: 'iframe';
   sandboxed?: boolean;
-  opaque?: boolean;
+  transparent?: boolean;
 }
 
 export interface Iframe extends FormComponentWithLabel {
   type: 'iframe';
   sandboxed: boolean;
-  opaque?: boolean;
+  transparent?: boolean;
 }
 
 const iframeFields = formComponentWithLabelFields.concat([
   FieldSchema.defaultedBoolean('sandboxed', true),
-  FieldSchema.defaultedBoolean('opaque', false)
+  FieldSchema.defaultedBoolean('transparent', true)
 ]);
 
 export const iframeSchema = StructureSchema.objOf(iframeFields);