+++ modules/tinymce/src/themes/silver/main/ts/ui/dialog/IFrame.ts
@@ -34,13 +34,13 @@ const getDynamicSource = (initialData: Optional<string>): IFrameSourcing => {
 
 const renderIFrame = (spec: IframeSpec, providersBackstage: UiFactoryBackstageProviders, initialData: Optional<string>) => {
   const isSandbox = spec.sandboxed;
-  const isOpaque = spec.opaque;
+  const isTransparent = spec.transparent;
+  const baseClass = 'tox-dialog__iframe';
 
   const attributes = {
     ...spec.label.map<{ title?: string }>((title) => ({ title })).getOr({}),
     ...initialData.map((html) => ({ srcdoc: html })).getOr({}),
     ...isSandbox ? { sandbox: 'allow-scripts allow-same-origin' } : { },
-    ...isOpaque ? { class: 'tox-dialog__iframe tox-dialog__iframe--opaque' } : { class: 'tox-dialog__iframe' }
   };
 
   const sourcing = getDynamicSource(initialData);
@@ -53,7 +53,8 @@ const renderIFrame = (spec: IframeSpec, providersBackstage: UiFactoryBackstagePr
       uid: newSpec.uid,
       dom: {
         tag: 'iframe',
-        attributes
+        attributes,
+        classes: (isTransparent ? [ baseClass ] : [ baseClass, baseClass + '--opaque' ])
       },
       behaviours: Behaviour.derive([
         Tabstopping.config({ }),